fix: issue with deplopy script

+++ deploy.sh
@@ -54,13 +54,18 @@ then
 fi
 
 echo "Copy files..."
-find $deployDir -type f -not -name 'CNAME' -delete
+find $deployDir -type f -not -name '.git' -not -name 'CNAME' -not -path '*/[@.]*' -delete
 cp -r dist/browsify/** $deployDir
 
-echo "Push to server..."
+echo "Push to deploy branch..."
 git --git-dir=$deployDir/.git --work-tree=$deployDir add .
 git --git-dir=$deployDir/.git --work-tree=$deployDir commit -m "Deploy v$newVersion"
 git --git-dir=$deployDir/.git --work-tree=$deployDir push origin $deployBranch
+if [ $? != 0 ]
+then
+    >&2 echo "Failed to push to deploy branch."
+    exit 3
+fi
 
 echo "Clean temporary repository..."
 rm -rf $deployDir
